Use AddOptionMenu instead of redefining Option menu
This is way better approach to how the things should be done.
As we bumped version requirement to 2.2.0 recently, I guess
it's not bad to raise it a little more. So now minimum engine
version is GZDoom 2.4.0, and we can do things how they should be
done, not like, 'setting up crutches and inventing bicycles'.

Also the menu is now on bottom of Options menu, rather than on top.